requirements.txt
python-dotenv
requests
tzdata
flask
psycopg2-binary
google-ads
google-auth-oauthlib
